requirements.txt
Django==5.0.0
djangorestframework==3.15.1
mysqlclient==2.2.4
python-decouple==3.8
markdown==3.6
django-filter==21.1